refactor(plugin-oci): rename oci_image to docker_build, split the module
The Dockerfile rule was called `oci_image`, but every attribute it has
names a docker concept — `dockerfile`, `context`, `build_args`, `stage`,
`builder`, `secrets`, `ssh`, `cache_from`, `cache_to` — and it depends on
a probe target whose only job is to ask buildx a question. It is a
`docker_build`, and calling it `oci_image` claimed the standards name for
the one driver that cannot uphold the standard's properties.

The name matters more than usual here because a daemonless, layer-based
image rule is coming, and that rule is the one users and agents should
reach for by default: no daemon, no QEMU for cross-arch, and a cache key
that actually covers what the build reads. It should be `oci_image`, and
that name has to be free.

None of this is released — `crates/plugin-oci` is not on master — so the
rename costs nothing today and would be permanent the day #159 merges.

Also splits `pluginoci/mod.rs`: the Dockerfile driver moves to its own
`docker_build.rs`, and `mod.rs` keeps only what more than one driver
needs (the archive-format enum, platform-string handling, the dep-list
helpers, the test fakes). `mod.rs` was 3353 lines hosting one driver plus
everything shared; adding a second image driver to it was not an option.

Drive-bys, both stale rather than new:
- the cdylib docs still described `skopeo` as a host capability, three
  commits after it stopped being one, and still said "four drivers".
- a test's doc comment had been orphaned onto the following test, so
  `run_names_the_fix_when_the_builder_has_no_file_exporter` carried a
  first line about `containerimage.digest`.
