+++ src/commands/bootstrap.rs
@@ -133,7 +133,7 @@ pub fn new_engine() -> anyhow::Result<(Arc<engine::Engine>, ShutdownTrigger)> {
     e.register_managed_driver(|_| Box::new(pluginhttp::Driver))?;
     // The `oci_*` drivers are not compiled in: like the go plugin, they ship as
     // a separate cdylib loaded from a `path:`/`url:` manifest entry
-    // (`heph-oci-plugin.json`), under their own `oci_image` / `oci_pull` /
+    // (`heph-oci-plugin.json`), under their own `docker_build` / `oci_pull` /
     // `oci_push` / `oci_load` names.
     e.register_managed_driver(|_| Box::new(pluginnix::Driver::new(home_dir.join("nix-driver"))))?;
